fixes #273
Fixes some bugs on edge cases in StrategyFromPaper and MinimumRectangleStrategy which got triggered by the FixedTreasure-Hider
